\name{average_specified_comparison}
\alias{average_specified_comparison}
\title{average_specified_comparison}
\usage{
average_specified_comparison(fit, df, input, low, high)
}
\arguments{
  \item{fit}{fitted glm object}

  \item{df}{data frame for evaluating predictions}

  \item{input}{string representing the name of the input
  feature you want an APC for}

  \item{low}{low value of input feature}

  \item{high}{high value of input feature}
}
\description{
Makes an average predictive comparison for two specified
values of the input variable of interest
}
\examples{
print("hi")
}

