\name{compute_apc_from_pairs}
\alias{compute_apc_from_pairs}
\title{compute_apc_from_pairs}
\usage{
compute_apc_from_pairs(predictionFunction, pairs, u, v, absolute = FALSE)
}
\description{
(abstracted this into a separate function from
\code{get_apc} so we can more easily do things like
\code{get_apc_with_absolute})
}

