\name{GetApcDF}
\alias{GetApcDF}
\title{GetApcDF}
\usage{
GetApcDF(predictionFunction, df, inputVars, ...)
}
\arguments{
  \item{predictionFunction}{}

  \item{df}{data frame with data}

  \item{inputVars}{inputs to the model}

  \item{...}{extra parguments passed to GetPairs used to
  control Weight function}
}
\description{
makes average predictive comparison for all specified
inputs
}

