\name{GetComparisonDFFromPairs}
\alias{GetComparisonDFFromPairs}
\title{GetComparisonDFFromPairs}
\usage{
GetComparisonDFFromPairs(predictionFunction, pairs, u, v)
}
\description{
(abstracted this into a separate function from
\code{GetApc} so we can more easily do things like
\code{GetApcWithAbsolute})
}

