\name{ComputeAPCFromPairs}
\alias{ComputeAPCFromPairs}
\title{ComputeAPCFromPairs}
\usage{
ComputeAPCFromPairs(predictionFunction, pairs, u, v, absolute = FALSE)
}
\description{
(abstracted this into a separate function from
\code{GetAPC} so we can more easily do things like
\code{GetAPCWithAbsolute})
}

