\name{PlotPairCumulativeWeights}
\alias{PlotPairCumulativeWeights}
\title{PlotPairCumulativeWeights}
\usage{
PlotPairCumulativeWeights(pairs, numOriginalRowNumbersToPlot = 20)
}
\description{
For a sample of transition start rows, we plot rank of
transition end (by increasing weight) vs. cumulative
weight. This gives a sense of how much weight is going into
the nearest points vs. further ones.
}

